game

+++ core/src/ru/samsung/gamestudio/GameSettings.java
@@ -12,8 +12,8 @@ public class GameSettings {
     public static final float SHIP_FORCE_RATIO = 10;
     public static final float TRASH_VELOCITY = 20;
     public static final float STARTING_TRASH_APPEARANCE_COOL_DOWN = 2000;
-    public static final float TRASH_WIDTH = 140;
-    public static final float TRASH_HEIGHT = 100;
+    public static final int TRASH_WIDTH = 140;
+    public static final int TRASH_HEIGHT = 100;
     public static final float SHOOTING_COOL_DOWN  = 1000;
     public static final float BULLET_VELOCITY = 200;
     public static final int BULLET_WIDTH = 15;
